Add an extra command to search DB clusters by given tags
<!--
A description of what this pull request does.
-->
There is a manual process to gather DB Clusters and VPC to ugprade in
latest versions of Postgres for example. Evolve DB factory to have a
command which can return the results in a table with the following
usage:

`dbfactory cluster -t 'DatabaseType=multitenant-rds'

or with multiple tags:

`dbfactory cluster -t 'DatabaseType=multitenant-rds' -t 'Owner=cloud-team'``

https://mattermost.atlassian.net/browse/MM-34230